DR - Save 24 bytes in the PSO initializer #rb none #rnx
[CL 4752987 by Rolando Caloca in Dev-Rendering branch]

+++ Engine/Source/Runtime/RHI/Private/PipelineStateCache.cpp
@@ -115,7 +115,7 @@ static void HandlePipelineCreationFailure(const FGraphicsPipelineStateInitialize
 	for(int32 i = 0; i < Init.RenderTargetFormats.Num(); ++i)
 	{
 		//#todo-mattc GetPixelFormatString is not available in scw. Need to move it so we can print more info here.
-		UE_LOG(LogRHI, Error, TEXT("0x%x"), Init.RenderTargetFormats[i]);
+		UE_LOG(LogRHI, Error, TEXT("0x%x"), (uint32)Init.RenderTargetFormats[i]);
 	}
 	
 	UE_LOG(LogRHI, Error, TEXT("Depth Stencil Format:"));
